server/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
fastapi-jwt-auth==0.5.0
fastapi-mail==1.2.4
pydantic[email]
pydantic==1.10.3